# users-service/src/main/java/com/users_service/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.users_service;

import com.users_service.NewUser;
import com.users_service.NewUserApi;
import com.users_service.UserUpdate;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-29T13:45:58.449049799+05:30[Asia/Kolkata]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "users", description = "the users API")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_USERS_ADD_POST = "/users/add";
    /**
     * POST /users/add : registers a new user
     *
     * @param newUserApi  (required)
     * @return OK (status code 200)
     *         or Server Error (status code 500)
     */
    @Operation(
        operationId = "usersAddPost",
        summary = "registers a new user",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "500", description = "Server Error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_USERS_ADD_POST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> usersAddPost(
        @Parameter(name = "NewUserApi", description = "", required = true) @Valid @RequestBody NewUserApi newUserApi
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_PUT = "/users/";
    /**
     * PUT /users/
     *
     * @param userUpdate  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Server Error (status code 500)
     */
    @Operation(
        operationId = "usersPut",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "404", description = "Not Found"),
            @ApiResponse(responseCode = "500", description = "Server Error")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = UsersApi.PATH_USERS_PUT,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> usersPut(
        @Parameter(name = "UserUpdate", description = "", required = true) @Valid @RequestBody UserUpdate userUpdate
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_REGISTER_POST = "/users/register";
    /**
     * POST /users/register : registers a new user
     *
     * @param newUser  (required)
     * @return OK (status code 200)
     *         or Server Error (status code 500)
     */
    @Operation(
        operationId = "usersRegisterPost",
        summary = "registers a new user",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "500", description = "Server Error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_USERS_REGISTER_POST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> usersRegisterPost(
        @Parameter(name = "NewUser", description = "", required = true) @Valid @RequestBody NewUser newUser
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_USER_ID_DELETE = "/users/{user_id}";
    /**
     * DELETE /users/{user_id} : delete a user based on user_id
     *
     * @param userId Numeric id of user that needs to be deleted (required)
     * @return OK (status code 200)
     *         or Not Found (status code 400)
     *         or Server Error (status code 500)
     */
    @Operation(
        operationId = "usersUserIdDelete",
        summary = "delete a user based on user_id",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "400", description = "Not Found"),
            @ApiResponse(responseCode = "500", description = "Server Error")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = UsersApi.PATH_USERS_USER_ID_DELETE
    )
    default ResponseEntity<Void> usersUserIdDelete(
        @Parameter(name = "user_id", description = "Numeric id of user that needs to be deleted", required = true, in = ParameterIn.PATH) @PathVariable("user_id") Integer userId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_USERS_USER_ID_GET = "/users/{user_id}";
    /**
     * GET /users/{user_id} : Get a user based on user_id
     *
     * @param userId Numeric id of user that needs to be fetched (required)
     * @return OK (status code 200)
     *         or Not Found (status code 400)
     *         or Server Error (status code 500)
     */
    @Operation(
        operationId = "usersUserIdGet",
        summary = "Get a user based on user_id",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "400", description = "Not Found"),
            @ApiResponse(responseCode = "500", description = "Server Error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_USERS_USER_ID_GET
    )
    default ResponseEntity<Void> usersUserIdGet(
        @Parameter(name = "user_id", description = "Numeric id of user that needs to be fetched", required = true, in = ParameterIn.PATH) @PathVariable("user_id") Integer userId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}